requirements.txt
Django==1.6
Pillow==2.3
South==0.8.4
django-boto==0.3
django-haystack==2.1
django-storages==1.1
lxml==3.3.1
requests==2.2
wsgiref==0.1
pysolr==3.2
nose==1.3
django-widget-tweaks==1.3
django-form-utils==1.0.1
django-tastypie==0.11